// src/types/enums/product.ts
export enum ProductionEnergySource {
    WIND = 'wind',
    SOLAR = 'solar',
    HYDRO = 'hydro',
  }
  
  export enum HydrogenType {
    GREEN = 'green',
    BLUE = 'blue',
    GRAY = 'gray',
  }
  
  export enum ProductionMethod {
    SMR = 'smr',
    ELECTROLYSIS = 'electrolysis',
  }
  
  export enum ApprovalStatus {
    PENDING = 'pending',
    APPROVED = 'approved',
    REJECTED = 'rejected',
  }
  
  export enum TransportationType {
    LAND = 'land',
    MARITIME = 'maritime',
  }